reconfigures epic web inngest structure and hook into the webhooks har har
