+++ apps/epic-web/.env.development
@@ -13,6 +13,7 @@ NEXTAUTH_SECRET=XKmGZcoWOTDEVSADasdsavJTa2Oyh8f4xUMV0=
 PORT=3021
 NEXTAUTH_URL=http://localhost:3021
 CLOUDINARY_VIDEO_BUCKET_NAME=epic-web
+INNGEST_APP_NAME=epic-web-next
 
 # Sanity
 NEXT_PUBLIC_SANITY_PROJECT_ID=i1a93n76
