+++ apps/epic-web/.env.production
@@ -12,6 +12,7 @@ NEXT_PUBLIC_SEO_KEYWORDS="web development, programming, javascript, full-stack, 
 NEXT_PUBLIC_PRODUCT_DESCRIPTION="Learn full-stack web development with Kent C. Dodds"
 CLOUDINARY_VIDEO_BUCKET_NAME=epic-web
 NEXT_PUBLIC_AMPLITUDE_API_KEY=6332ef8ebd7a3b750221c9c3bcc6cd9a
+INNGEST_APP_NAME=epic-web-next
 
 EMAIL_SERVER_HOST=smtp.postmarkapp.com
 EMAIL_SERVER_PORT=587
