+++ packages/inngest/src/index.ts
@@ -1,2 +1,3 @@
 export * from './inngest.server'
 export * from './convertkit.inngest'
+export * from './events'
